BorderlandsGOTYEnhancedFix: Initial commit of working fix for ultrawide resolutions and FOV
Like the title mentions inital commit that adds
support for ultrawide resolutions and fixes FOV
where it dynamically scales with resolution as the
game employes vert- scaling. The FOV fix emulates
what hor+ would look and feel like.

Two outstanding issues remain that will hopefully
be addressed in the future either by me or someone
else. There is strange graphics artifacting on the
left side of the screen when resolution exceeds
21:9. The artifacting mirrors the graphics
presented on the right side of the screen. This
may be an Unreal Engine 3 limitation. Hard to say
if this can be fixed. The other issue is that the
HUD is streched on resolutions exceeding 16:9
there should be some knob in the binary to keep
the HUD constrained to 16:9, but that is to be
seen.

	new file:   .gitignore
	new file:   .gitmodules
	new file:   CMakeLists.txt
	modified:   LICENSE
	modified:   README.md
	new file:   inc/utils.hpp
	new file:   install.ps1
	new file:   safetyhook
	new file:   spdlog
	new file:   src/main.cpp
	new file:   src/utils.cpp
	new file:   yaml-cpp
	new file:   zydis