FOV fix: Working but not optimal FOV fix.
The previous FOV fix was a placebo because the FOV
was cranked to the max which distorted if the fix
was working or not. Spoilers it wasnt.
This commit adds the FOV fix. The fix is not
optimal and would be much better to just force
hor+ scaling if I can find it.
This current fix although fixes the FOV also takes
away any gameplay dynamics like widening of the
camera during running, the initial bounce of the
camera when loading in for the first time when
starting the game, and probably more that I havent
seen yet.

	modified:   src/main.cpp

Uploaded a better image showcasing the fix and
updated README.md to show correct image.

	modified:   README.md
	modified:   images/BorderlandsGOTYEnhancedFix_1.gif