BorderlandsGOTYEnhancedFix: Massive overhaul
For sanity I added back every memory location that
the game writes the resolution to from the
launcher. This was done for pure sanity to make
sure all bases where covered in case something
goes wrong and resolution cant be blamed.

LOG file will now have compiler information to
assist with debugging.

	modified:   src/main.cpp

Added new function which will get compiler info,
this is to be used to help with debugging.

	modified:   src/utils.cpp
	modified:   inc/utils.hpp

Release builds for Github shall now be made with
GCC and will staticly link libgcc and libstdc++.
Using GCC and those libs will prevent user errors
with MSVC where certain DLLs are missing which
will cause the fix to not load and subsequently
not load the game. This has been made into a
cmake build option, those wishing to build from
source can still use any compiler and any build
system, but for release builds on Github GCC
builds will be forced to prevent missing DLL
issues as stated previously.

Also made the binaries go into their own folder
as opposed to the build folder with other cmake
artifacts for convenience.

	modified:   CMakeLists.txt

Added bin folder to ignore list so binaries are
not accidently uploaded Github repo.

	modified:   .gitignore

Changed where the script will attempt
to look for dll file. Note it will recursively
traverse the bin folder and use the first dll it
finds regardless if its the debug or release dll.
This is by design and NOT a bug.

	modified:   install.ps1