BorderlandsGOTYEnhancedFix: Improved FOV fix, fixed artifacting in resolutions > 21:9
Removed all things GCC and went back to MSVC as the default build
platform as I figured out how to link in all dependencies into the final
DLL so users are not prompted that they are missing libraries and files
on thier computers.

	modified:   CMakeLists.txt

Updated documentation. Got rid of the GCC section as that is now null
info and no longer applies. Also updates the known issues chapter of the
readme. Crossed out when certain issues where fixed and with what
version and also marked the HUD problem, which is really a none problem
in the grand scheme of things that it shall not be fixed.

	modified:   README.md

Added comment to make sure users max out in game FOV to 120.

	modified:   install.ps1

Added code that improved the FOV fix and additional hook that fixes the
artifacting that occurs on the left side of the screen when resolution
exceeds 21:9.
For FOV we now scale with the ingame FOV expected value, it is assumed
that the ingame FOV has been set to 120 which is the max. We use this as
the base line as the game knows its limits of when the FOV in game will
break the game and to assure it doesnt break the fix we need this to be
maxed out. The new fix will now take the user specified FOV in the
config file and scale it with what the game calculates the current FOV
should be, this allows us to mimic what would be happening if the game
scaled hor+ instead of vert-.
For the artifacting this is a fix I cannot explain. I did not delve
at all into the code as to why just doing +1 to the height at this place
in the code will get rid of the artifacting, but as stated in the
documentation in the source file I can only assume we are now either
passing some sort of check or failing some sort of check which allows
the game to render stuff properly hence the artifacting is gone.

	modified:   src/main.cpp