feature/scaleSprintFov: Add feature proposed by Issue #3
Original feature request asked for the ability to clamp FOV when
sprinting. The reason for this request was that at high FOV values the
game would really pull back the camera causing an undesirable fisheye
effect.

This solution does not clamp the FOV at some value as that would
interfere with gameplay where the game would hit the clamp very quickly
and it would look weird and awkward and when sprinting would stop it
would take a visible moment for the camera to regress to default
position.

To address this the feature allows the user to scale the sprinting FOV,
by default it will be disabled and be set to 1.0. The user must enable
the feature themselves and set the appropriate scaling factor themselves
given that FOV is an exponential thing, ie the lower the FOV goes the
less and less visible change will be observed and the higher the FOV
goes the more pronounced further FOV changes will be made. For example
if by default the FOV is 50, going down to 40 will not be as visibly
drastic as going 50 to 60. The same applies for this fix the lower you
go the less pronounced each step will be and vice versa.

New feature so bumped version to v2.1.0

	modified:   src/main.cpp

Added feature construct to yaml file generation

	modified:   install.ps1